Expand field element tests
- Add `impl_field_identity_tests!`
- Use `impl_field_sqrt_tests!`

NOTE: this change uncovered some test failures with `p256::Scalar::sqrt`
and `p384::Scalar::sqrt` implementations, tracked as #789.

There are now TODOs in the relevant places in the code.
